+++ stories/Combo/component/create-combo.tsx
@@ -1,0 +1,259 @@
+import React, { useEffect } from 'react';
+import G6 from '../../../src';
+import { IGraph } from '../../../src/interface/graph';
+import { ComboConfig, GraphData } from '../../../src/types';
+import { IGroup } from '@antv/g-base/lib/interfaces';
+
+let graph: IGraph = null;
+
+const data: GraphData = {
+  nodes: [
+    {
+      id: '1',
+      dataType: 'alps',
+      name: 'alps_file1',
+      conf: [
+        {
+          label: 'conf',
+          value: 'pai_graph.conf',
+        },
+        {
+          label: 'dot',
+          value: 'pai_graph.dot',
+        },
+        {
+          label: 'init',
+          value: 'init.rc',
+        },
+      ],
+    },
+    {
+      id: '2',
+      dataType: 'alps',
+      name: 'alps_file2',
+      conf: [
+        {
+          label: 'conf',
+          value: 'pai_graph.conf',
+        },
+        {
+          label: 'dot',
+          value: 'pai_graph.dot',
+        },
+        {
+          label: 'init',
+          value: 'init.rc',
+        },
+      ],
+    },
+    {
+      id: '3',
+      dataType: 'alps',
+      name: 'alps_file3',
+      conf: [
+        {
+          label: 'conf',
+          value: 'pai_graph.conf',
+        },
+        {
+          label: 'dot',
+          value: 'pai_graph.dot',
+        },
+        {
+          label: 'init',
+          value: 'init.rc',
+        },
+      ],
+    },
+    {
+      id: '4',
+      dataType: 'sql',
+      name: 'sql_file1',
+      conf: [
+        {
+          label: 'conf',
+          value: 'pai_graph.conf',
+        },
+        {
+          label: 'dot',
+          value: 'pai_graph.dot',
+        },
+        {
+          label: 'init',
+          value: 'init.rc',
+        },
+      ],
+    },
+    {
+      id: '5',
+      dataType: 'sql',
+      name: 'sql_file2',
+      conf: [
+        {
+          label: 'conf',
+          value: 'pai_graph.conf',
+        },
+        {
+          label: 'dot',
+          value: 'pai_graph.dot',
+        },
+        {
+          label: 'init',
+          value: 'init.rc',
+        },
+      ],
+    },
+    {
+      id: '6',
+      dataType: 'feature_etl',
+      name: 'feature_etl_1',
+      conf: [
+        {
+          label: 'conf',
+          value: 'pai_graph.conf',
+        },
+        {
+          label: 'dot',
+          value: 'pai_graph.dot',
+        },
+        {
+          label: 'init',
+          value: 'init.rc',
+        },
+      ],
+      comboId: 'combo1',
+    },
+    {
+      id: '7',
+      dataType: 'feature_etl',
+      name: 'feature_etl_1',
+      conf: [
+        {
+          label: 'conf',
+          value: 'pai_graph.conf',
+        },
+        {
+          label: 'dot',
+          value: 'pai_graph.dot',
+        },
+        {
+          label: 'init',
+          value: 'init.rc',
+        },
+      ],
+      // comboId: 'combo2',
+    },
+    {
+      id: '8',
+      dataType: 'feature_extractor',
+      name: 'feature_extractor',
+      conf: [
+        {
+          label: 'conf',
+          value: 'pai_graph.conf',
+        },
+        {
+          label: 'dot',
+          value: 'pai_graph.dot',
+        },
+        {
+          label: 'init',
+          value: 'init.rc',
+        },
+      ],
+      // comboId: 'combo2',
+    },
+  ],
+  combos: [
+    {
+      id: 'combo1',
+      label: 'Combo-1',
+    },
+    // {
+    //   id: 'combo2',
+    //   label: 'Combo-2',
+    //   parentId: 'combo1'
+    // },
+  ],
+  edges: [
+    {
+      source: '1',
+      target: '2',
+    },
+    {
+      source: '1',
+      target: '3',
+    },
+    {
+      source: '2',
+      target: '4',
+    },
+    {
+      source: '3',
+      target: '4',
+    },
+    {
+      source: '4',
+      target: '5',
+    },
+    {
+      source: '5',
+      target: '6',
+    },
+    {
+      source: '6',
+      target: '7',
+    },
+    {
+      source: '6',
+      target: '8',
+    },
+  ],
+};
+
+const CreateCombo = () => {
+  const container = React.useRef();
+  useEffect(() => {
+    if (!graph) {
+      graph = new G6.Graph({
+        container: container.current as string | HTMLElement,
+        width: 1000,
+        height: 800,
+        groupByTypes: false,
+        modes: {
+          default: [
+            'drag-canvas',
+            'zoom-canvas',
+            'drag-node',
+            'drag-combo'
+          ],
+        },
+        fitView: true,
+        defaultCombo: {
+          type: 'circle'
+        },
+        layout: {
+          type: 'dagre',
+        },
+      });
+
+    }
+    graph.data(data);
+    graph.render();
+    graph.on('canvas:click', e => {
+      graph.createCombo({
+        id: 'combo2',
+        label: 'Combo-2',
+        parentId: 'combo1',
+      }, ['7', '8']);
+      setTimeout(() => {
+        graph.updateCombo('combo2');
+      }, 500)
+    });
+
+  });
+  return <div ref={container}>
+  </div>;
+};
+
+export default CreateCombo;